+++ app/components/analytics/index.ts
@@ -1,3 +1,4 @@
 export { default as AnalyticsChartCard } from "./AnalyticsChartCard";
 export { default as AnalyticsTotalsRow } from "./AnalyticsTotalsRow";
 export { default as ChartTooltip } from "./ChartTooltip";
+export { default as MonthYearPicker } from "./MonthYearPicker";
